+++ src/main/java/com/cinemania/model/PopularInTheatreMovies.java
@@ -20,6 +20,6 @@ public class PopularInTheatreMovies {
 	private Long movieId;
 	private String movieName;
 	private String movieImgUrl;
-	private int movieRating;
+	private float movieRating;
 	
 }
